Create codemod to invoke update as a script
